Describing what the project teaches
I added a readme to specify what the project is inteded to teach me.
I think I may update it after to describe how it went.
I may also try to be more abstract later on, not focus so much on
specifics unless the entire exercise is about a single specific thing.